split test_viernulvier.py into smaller files
